Adding `Command::Success`.
Two of quickwit actors (the indexer and the merge planner)
have a special message to instruct them to exit and success.

This change makes it into an actual command.

The aim is not just to factorize the logic, it is also to make
it possible for the ParallelActor (in an upcoming PR) to properly
react to this command.