Bugfix: Stopping the merge planner cannot rely on healthcheck
The mechanic that decide whether the merge planner should be
stopped is broken.

Indeed, it relies on the idea that the packager healthcheck is faulty
while the merge_planner is healthy and running, but these calls
happen right after the overall healthcheck has been called.

The progress is therefore always reset at this point and the condition
never triggers.