Introduce an independant MergePublisher.
As explained in #668, the fact that the publisher keeps processing
pre-upload request in order makes it so that merge split upload
can stall the regular indexing path, and affect time to search.

This PR changes the graph topology by introducing
a merge publisher.

As an unexpected benefit it simplifies the termination logic.
Now the cycles only exists on the merge pipeline
(merge planner -> ... -> merge publisher)

Therefore it is relatively straightforward to stop the merge planner
as soon as the indexing publisher terminates.

This change also fixes the logging in the publisher.

Closes #668.