Adding automated tag pruning logic
A predicate is extracted from the user in the form of an AST.
This predicate is guaranteed to be true for all doc matching the
initial query.

This predicate is expanded using the rules described in #887.
Likewise, in the packager, the tags are expanded using the
same rules

In the PostgreSQL code, the predicate is expanded into a SQL
expression. Other metastore may rely on direct predicate
evaluation algorithm, over the AST.

Co-authored-by: Ryad ZENINE <r.zenine@gmail.com>
Co-authored-by: Paul Masurel <paul@quickwit.io>