Fixing the coverage CI.
This is a tricky problem.
In the current setting, localstack should be accessed via localhost,
not localstack.

This is because right now the CI tests are running on a host,
not in a dockerized environment.
(The services are running in containers)

I am not entirely sure how this is used to work.