Refactoring of quickwit-actors
This refactoring includes different changes:

- Instead of having a Message associated type, actor now can
have receive different Message by implementing a `Handler`
trait.
One hidden benefit is that some of the message can be private.
(e.g. loop messages).

- Message can yield a typed reply. The type is an associated type
of the Handler trait.

- There are no more Sync and Async actors. All handlers require
to implement an `async` method. What used to be SyncActors need to
specify that want to be executed on an independent runtime, on a
dedicated thread, usign the `Actor::runner` method.

The refactoring was mainly motivated by the introduction of services in
quickwit: the IndexingService (currently named IndexingServer),
and the upcoming PushAPIService.