Refactoring the doc mapper.
The current implementation is relying on a struct with the union of all
of the field options, and some dynamic logic to ensure that the enterred
options are all there and correct.

Functionally, there are three modifications:
- it forbids `array<bytes>` as tantivy does not support
array<bytes> + fastfield.
- some of the error message are relying on `serde`,
and are therefore different. One class of error message have some
regression: if `tokenizer` is added to an int field, we detect and
return a valid error, but we do not mention that this option is only
valid for TextOptions any more.
- More consistent error format. We always report the faulty field.