Fix how quickwit handles "." in field names.
After this PR we still accept field names with a ".".
On the query side this dot then needs to be escaped with a backslash
to disambiguate it from the "." of Json fields, and the "." of
object fields.

Under the hood, we still rely on tantivy's query parser.
The hack consists in escaping the tantivy field name itself.

This PR updates tantivy in order to include
https://github.com/quickwit-oss/tantivy/pull/1356.

Closes #1338
Closes #1334