Add a search service to address REST requests on indexer only nodes.
We add a search service to address REST requests in the UI of index-only
nodes as quickfix for #1482.
It will only address the root part of the requests.

That search service is not mounted as a gRPC service and hence will not address
leaf requests.

We also return an explicit error if there are no search nodes in the cluster.
(before we were panicking).

Closes #1482