Solves a race condition at the end of an actor.
The actor was updating its state before submitting
its ActorExitStatus. The healthcheck was then interpreting
the lack of exit status as the result of a panic, when it
was just a race condition.

This PR changes the actor state so that it includes two
exit states: success and failure.

This PR also simplify the code associated to the join handle
and the reporting of the exit status.

We do not rely on a separate watch channel anymore.
Also the JoinHandle code does not allow joining several time
anymore, which simplifies the code.

Closes #1514