Factorized the http client and the credential provider used in rusoto.
This solve a bug observed in CI.
Since we upgrade rusoto, by default, its http client does not allow
http (as opposed to https) connections.

This was spotted and fixed for the S3 storage unix test, but the problem
is still there for Kinesis.

This PR makes sure that both kinesis and s3 actually use the same code
to create their http client.

It also shares the logic used to get credentials.