Introducing a sequencer actor.
File uploads are done concurrently.
They may end in a different order than they started.

Before this PR, maintaining the ordering was the job of the publisher.
After this PR, this sequencing logic is done by a dedicated actor
called Sequencer.

The purpose of this PR is to pave the way to make the Publisher a
singleton Service/Actor (as opposed to one actor per index).

We do not want to observe orders for all indices.
One slow upload on one index does not need to delay the publication of
split for all indices. It was therefore necessary to keep the sequencing
within the per-index indexing pipeline.

Besides, it separate the concern, and makes the unit test & logic much
cleaner.