Change in RocksDB configuration to avoid disk usage inflation.
We are using one column family per index and a shared WAL for
all column family.

RocksDB seem to not being able to truncate the WAL itself
on delete range unfortunately, and the default policy tends
to be per-cf which leads to a very high disk usage
when using a lot of indexes.

This PR force flush once the WAL reaches 100MB.
This is quite low, to avoid the push service to block
for too long.