Failpoint test fix.
After removing the use of the Kill message,
some actors could be stuck awaiting for a message, and never
checking for possible message.

Relying on a Kill message is not an option, due to the
recycling of mailbox introduced in the supervisor PR.

The fix consists in adding a Nudge message that
is No-op.