Simplifies and fixes the local split cache.
After the fix, there is only one split cache shared
by all of the indexes.

Splits are all stored without any specific hierarchy in the
`data_dir/cache` directory.

The split cache evince oldest splits first, and introduces
an extra constraint on the creation date to make sure
dangling splits eventually get removed even when the split is not under
capacity pressure.

---

The Local Split Store is a local cache used to improve the performance of indexing nodes.
Its purpose is simple: when a new split is freshly created, it is usely merged
very rapidly after.

In order to prevent this merge to force its download, we store it in the
`LocalSplitStore`. This store is just a cache: a cache miss is acceptable and
just means that the split will be redownloaded.

The local split store eviction policy however, is rather uncommon.
On our happy path, a split is stored into the cache, and is then used only once
to undergo a merge.

For this reason, we simply offer a way to `move splits into the cache`,
and `move splits out of the cache`. A split is removed from the split store
after its first access.

Of course a failed merge could require accessing a given split more than once. In that
case the split will be downloaded again.

The cache size is limited by 3 things:
- a maximum number of splits as defined in the `SplitStoreQuota`.
- a maximum number of bytes as defined in the `SplitStoreQuota`.
- finally, we evince older splits to make sure that newest split and the oldest
split only differ by at most `SPLIT_MAX_AGE`.

The point of this final rule invariant is to make sure that the disk space will be
if the cache is NOT under pressure but some splits are actually useless.

When adding a new split into the cache, if adding the split would break one of the following
limit, we simply remove split one by one starting by the oldest first, until the split
can be added.

Closes #2022