Added generic IO controls.
The `IoControls` object makes it possible to
- limit the throughput of Write and AsyncWrite object,
- abort a write if the killswitch is activated,
- record progress when performing IO
- record the number of writes write in a Prometheus counter.

In this PR, the IoControls object is plugged into the
the indexer, the merger and the delete pipeline.

The merge and the delete pipeline can optionally get a
separate write throughput limit, as defined in the index
configuration.

In the merge and the delete pipeline, split downloads and
actual index writing share the same throughput limit, and
attempt to consume it in a concurrent manner.

We rely on async-speed-limit for the throttling.
We could not use async_speed_limit::Resource directly as
it is not compatible with tokio.

This PR also does some minor refactoring on the merge pipeline
parameters.