Rationalization of IndexConfig/IndexMetadata.
This PR removes the redundancy between IndexConfig and IndexMetadata and
makes SourceConfig and IndexConfig versionable, and with backward
compatilibity.

This PR breaks backward compatibility in the metastore, as it expects a
version string tag on `SourceConfig`.

The current PR will break backward configuration.
It is not a necessity, but it requires a bit of effort to keep
backward compatibility.

Related to #2048