Fixes Bug in FileStorage
Writing in the file storage was not atomic.

So if some other process was polling the file storage
for instance, it could end up reading some version
of the meta file that was partially written, and therefore invalid JSON.

This is the bug that was observed in #2304.

This PR also prevents the file metastore from reloading
the data due to the polling loop, if it is the process acactually writing the metastore.

Closes #2304

Persisting things in an atomic way in the FileStorage.