Random `partition_id` for the ingest API source.
This PR changes the partition_id used in the ingest API.
After this changes, we use a randomly generated partition_ID,
and save it in RocksDB.

This solves two problems that we have.

We can then have two nodes running the ingest_api
as a duct tape distributed model. That way, their
checkpoint won't interfere.

Closes #2291

It also makes it possible to behave less poorly if the
rocksdb data dir is lost (for instance, when using a
ephemeral volume).

Before the change, the ingest API would add
give the new message auto increment positions starting at
0.
The source checkpoint in the metastore would prevent quicwit
from indexing any document until the previous position
is reached.

Closes #2310