Fixes the merge planner.
This PR tweaks the way the merge planner schedules its merge operation
in a subtle way to avoid adding a lot of large merge operations at the
same time... Hence stalling smaller, with a larger bang-for-the-buck
merge operations.

This is done by stopping queue merge operations as soon as push back is
experienced.

Closes #2348