No modulo in the partitioning logic.
The default value of max_num_partitions is now 200.
It is required, but users can set it to 2^32 if they want a "die before mixing partitions" behavior.
If the max_num_partitions is exceeded, we log a warning with the number of max_num_partitions, but also the number of docs in the workbench when it happens.

Closes #2458.