Remove mold usage in CI.
The current CI code is referring to the latest version
of a binary build on a mold's creator.
That's a recipe for breaking CI, exposing ourselves to change of API
(unlikely for a linker but well), supply chain attack etc, change
of licenses (for CI that's ok but for building our binares that would
have been a disaster).

Let's stick to the stock linker.