Large actor scheduler refactoring
The scheduler is now not an actor like in the actor framework,
but an adhoc "lighter" actor. It avoids the confusion of
having to create an actor with a fake scheduler mailbox etc.

Simulating jumping in time is now behaving in a smarter way:
We jump in time only if no other actors is processing messages.

The retry logic now works on a mock time to accelerate the unit test.