feature: prompt confirmation for `delete index`
The CLI interface of quickwit has a subcommand for deleting an index, given
it's ID. The subcommand, except for dry-run, has no safety built-in. I
thought about putting a `--yes` flag, like other subcommands in the CLI,
but I feared that this may be disrupting for existing client's
pipelines. This seems a more sensible choice because it doesn't
interfere with existing processes in place and it can be used only where
is necessary.