Large actor scheduler refactoring
The scheduler is not a actor in the sense
of the actor framework anymore.

It removes the necessity to create some fake
scheduler mailbox to spawn the scheduler
itself.

The scheduler also now has an improved logic to
simulate time shift.

It only jumps forward when no actors has any work to
do. Provided all of the processing is done in actor,
the results should be rigorously the same as if someone
used time::sleep... Only faster.