Update doc sections describing indexing behavior
This commit removes the section describing behavior with fields not
defined in the config since it is obsolete and the new behavior is
already covered in details in the previous section on the mapping
mode.

It also clarifies the section describing behavior with missing
fields, which currently doesn't mention that missing fast fields
will cause the entire record to be rejected.

Fixes #2711