Bugfix: Actor's mailbox content not dropped when the last Receiver is dropped.
This is due to a strange spec in flume. In flume, the last Receiver does
not cause the content in the channel to be dropped.
For this to happen, all Receiver AND Sender needs to be dropped.

This caused a bug in which an actor was stuck waiting for the result of
a `.ask(..)` message.

This PR explicitly drains mailbox on the drop of their underlying
channel.