Addressed a bad interaction between rustfmt and codegeneration.
With the addition of services codegeneration, we end up
creating new files that are not formatted in the sense of rustfmt.

As a result, every new run of `make fmt` results in a diff.
Currently, the problem is solved by commit the formatted version of the
code.

This is still not great, because now every call to `cargo build` shows a
diff until we run `make fmt`.

Finally, the current code does not make it clear enough what is
generated and what is not. The last thing we want is for people to start
editing/committing modified generated code.

This PR introduces the following solution:

- generated code stays in src/. Putting it in target is a common way to
fix these problem but it tends to confuse source code editor
GoToDefinition. If generated code is not something we want to write, we
often want to read it.

- generated code goes into codegen/ subdirectories. That subdirectory
is not covered by rustfmt anymore. We commit non formatted version of
the file (it does not hurt readability).