Add an option to force and wait for commit
Adds an option to force a commit or wait for the next commit while
issuing an ingest command. This commit only implements it for the
ingest endpoint. If all records in the batch fail, the command will
hang until at least one good document is indexed until #2825 is merged
in or the document parsing is moved to the early stages.

See #2699