Differentiates absence/presence of quotation marks in user query.
Before this change, a user text yielding more than one token
would be always interpreted as a phrase query.

jobtitle:editor-in-chief

Would therefore fail if the field was indexed tokenized but
without positions.

After this change
jobtitle:"editor-in-chief" is interpreted as a phrase query.
jobtitle:'editor-in-chief' is tokenized with the field tokenizer,
but interpreted as an intersection.
jobtitle:editor-in-chief is tokenized and interpreted as a phrase
query if positions are available for the field at the scale of the
split, and will return an error if it is not available.

Closes #3364