Ack pipeline termination and respawn faster.
This technically fixes two bugs:
We increased the quickwit default heartbeat duration to 30s.
This was done to avoid false positive when trying to detect
an actor being stuck (and not displaying any progress).

However, the supervision loop was not just checking for
the actor's progress.

It was also checking whether they were terminated or not.
In addition the heartbeat duration was used to determine the first
cooldown period before respawn. 30s seems very bad here.

This PR changes the code, merging the observe and the supervision
code.
Healthcheck only checks for progress when heartbeat has elapsed.
The rest of the loop iteration, we only check for the actor
status.

It also changes the time of the first respawn to 1s, independently
from Quickwit's heartbeat.

This PR also removes the shared Observe message as it was
an important footgun: in the indexing/merge pipeline, the Observe
message was a looptoken for the observe loop:
ny client calling Observe would spawn a different loop unknowingly.

The IndexingService was using it as a way to get observation.
That observation was however not updating the actor last observed state.

I instead exposed the quickwit actor crate Observe message that "does
the right thing".

Fixes #3699