Support but ignoring `lenient` property in MultiMatchQuery in the ES API.
Elasticsearch has a lenient property that just avoids
returning an error when facing a type mismatch between the query and the
schema.

For instance, searching a text in a integer field.

This PR does not add this feature. This definition of lenient is
actually already Quickwit's default behavior.
This PR however makes it possible to accept the "lenient" in request
bodies.

The property value is always ignored and we always behave as lenient.