Prevents `Observe` message emitted by Supervisors from stacking up.
This introduce a refresh observation method, that does not wait
for anything but just triggers an eventual refresh of the last
observation.

This method includes debouncing logic to avoid stacking up messages.

We then use it in supervisors, which the other merit of making the
Supervise handler return much faster.

Before we were waiting for the observation to timeout.

Closes #3485