Changes the last modified timestamp of newly reported splits. (#3867)
This is important to improve the hit rate of the split cache.
Before the PR, we used a last modified time of 0.

As a result, once the cache reached saturation, a newly reported
split could trigger the eviction of an old split that got accessed
once.

In other words, cache population could only happen when splits got
accessed. This is really a bad situation because the split cache
does not get any hint when splits are deleted. It only relies
on eviction to release space.

After this workaround, splits that are not accessed ends
up having a lower priority than new splits after
10mn without any access.