Fix the cluster change stream when nodes rejoin the cluster quickly
The issue is that the cluster logic keeps track of nodes joining and leaving
using the tuple `(node_id, incarnation_id)`, while client pools keep track of
nodes using only `node_id`.

When a node restarts and becomes ready quickly, the new node
`(node_id + incarnation_id + 1)` joins the cluster before the previous node
leaves it and the following cluster change events are triggered:

1. Add: `(node_id + incarnation_id + 1)`
2. Remove: `(node_id + incarnation_id)`

On the other hand, it translates at the pool level to:
1. Add: `node_id`
2. Remove: `node_id`

The bug used to be less visible because we used to diff the whole pool for
each Chitchat event, but now the diff computed incrementally.

To fix the issue, we only emit a `Remove` event when incarnation IDs match.
Additionally, when a node rejoins the cluster, instead of doing nothing, we
virtually emit two consecutive `Remove` and then `Add` events to give the pools
a chance to react to configuration changes.