Refactoring of the control plane.
This paves the way for the construction of a partial view of the
metastore, by isolating the model nested in the ingest controller as
a control plane model.

It also change the error handling. Now upon a metastore error
for which we do not know whether the control plane has lost consistency
or not, we kill the control plane.

Finally the control plane is now supervized.