Refactoring introducing the Cpu capacity.
This new CpuCapacity type is used to both expressed
the available capacity in indexer nodes, and
the load associated on shards.

As a result, the quantity we use for the latter
is semantically shifted from the CPU spent on the
indexer actor, to the estimated CPU cost of an
entire pipeline. (x4)