Introducing pipeline uid in indexing tasks/ indexing pipelines.
The scheduling algorithm is for the most part unchanged.
Only the allocation of shard to pipelines is modified.

Currently the logic yields similar results as before. In particular, the scheduling algorithm
may in some case, move shard from one pipeline to another.
In the future, we will prevent this behavior and instead close/open new shards.

Closes #4191