Increase the gRPC payload limits from 10MiB to 20MiB.
By default our REST client sends batch of 10MiB that gets expanded to
a little bit more after the lines are identified and the batch is
converted to a different format.

If the ingest request is sent to a searcher (like when we use helm with
an ingress pointing to a searcher), it has to be forwarded to an
ingester via gRPC.

This solution is imperfect as the rest payload is actually configurable.