Remove shards from  ingesters on source/index deletion.
The PR works by adding a RPC on ingester, to "retain shards":
the control plane can send a list of shards that it expects to see
on an indexer and the ingester deletes the shards that are not in that
list.

The control plane performs this RPC when a source is deleted,
or what an index is deleted, or upon restart.

This PR is incomplete and needs some work as described in #4274.
The "retain" mechanics was preferred over sending the list of shard
to remove precisely in prevision of #4274.

Also
- Bugfix in num shards metrics (we were returning the number of source)
- Adding a ingester -> shard table in the control plane model

Closes #4244.

Adding ingester list of nodes in shardtable model.

Added happy path to delete shars on delete source

Added unit test