First stab at bugfix.
We avoid evicting from the ingester pool a node that experienced a timeout or a transport error.

No code path would re-add it to the ingester pool, and that ingester pool is also used by the ingest source
resulting in bug #4336.

Instead we simply treat these errors as any other transient errors, and rely on the round robbin logic to avoid picking the same node over and over again.