Initializing the shard positions actor with the existing cluster state (#4621)
* Ensuring we don't lose the first few position updates.

We had two problems here.
a) the ingester loads in an asynchronous fashion, as it needs to replay
the WAL entirely.
Subscribing to shard positions events is part of this initialization.
We could be missing all of the event received before the initialization (via chitchat).
b) we start the cluster first. Long after we in the initialization
process, we start the ShardPositionsService. We could possibly be
missing the keys that were shared in between.

Closes #4622

* Apply suggestions from code review

Co-authored-by: Adrien "Code Monkey" Guillo <adrien@quickwit.io>

---------

Co-authored-by: Adrien "Code Monkey" Guillo <adrien@quickwit.io>