Fix race condition related to deboucing.
In Ingest v1, queues are created on apply plans.
This means that right after creating an index, documents cannot be
inserted.

This was true before, but was dependent on a race condition. It
became more apparent after we added debouncing of control plane apply plans.

This PR adds code to make it possible for caller of debounced rebuild
plan code, to get notified and wait for a rebuild to be done and
applied.