multi index leaf search request
execute one request per node, instead one request per index.
A leaf can receive now request over multiple indices.

Especially in cases with search requests on many indices this saves a
lot of requests and therefore memory. It also allows better control of
the memory consumption of a requests on a node, if it is not split up
over multiple requests.