introduce IntermediateLeafResult
LeafSearchResponse includes serialized aggregations to send them between
nodes. This is used also on a leaf, which causes the results to be
serialized and then deserialized for merging again _per_ split.

This PR introduces IntermediateLeafResult to be used instead.
It also adds a field `aggregation_type` to LeafSearchResponse to be able
to convert self contained between IntermediateLeafResult and
LeafSearchResponse.