multi index leaf requests (#4962)
* multi index leaf search request

execute one request per node, instead one request per index.
A leaf can receive now request over multiple indices.

Especially in cases with search requests on many indices this saves a
lot of requests and therefore memory. It also allows better control of
the memory consumption of a requests on a node, if it is not split up
over multiple requests.

* introduce IntermediateLeafResult

LeafSearchResponse includes serialized aggregations to send them between
nodes. This is used also on a leaf, which causes the results to be
serialized and then deserialized for merging again _per_ split.

This PR introduces IntermediateLeafResult to be used instead.
It also adds a field `aggregation_type` to LeafSearchResponse to be able
to convert self contained between IntermediateLeafResult and
LeafSearchResponse.

* multiple incremental collectors, pass agg limits

* Revert "introduce IntermediateLeafResult"

This reverts commit c968e4da6cc9a8bb5a3693c694005e1c99412d2a.

* fmt, cleanup

* remove outer incremental collector

* cleanup

* fix test

* fix test

* cleanup

* minor fixes

* clippy