Fixes control plane/metastore inconsistency due to open shard.
On scale up, rebalance, and get_or_open_shards, the control plane was:
- recording a shard on the metastore
- writing the shard
- record the shard on the control plane model.

Error handling was not done, so a transport failure on the metastore, or
(more likely) an error failure on init would break consistency between
the metastore and the control plane.

This PR factorizes the idea of opening a new shard for scale up,
rebalance and get_or_open_shards.

The new factorize logic goes:
- init first
- record in the metastore
- record the shard on the control plane model.

The last two steps are also factorized together to emphasize that we
keep the control plane and metastore in sync.

It works by forcing a restart of the control plane if the metastore
returns an error for which we don't know if the write was a success or
not.

Closes #5008
Closes #5020
Closes #5013