Add missing multi match modes to ES compatible Query API
According to [doc][1] multi_match has 6 modes. Best fields and Cross
fields can be converted to Most fields as they close by logic and I
couldn't find their implementations.

BoolPrefix was implemented and I added it to the parser.

Tests was extended to cover all cases, for that I changed a bit of
structures and added missing traits implementations.

[1]: https://opensearch.org/docs/latest/query-dsl/full-text/multi-match/#multi-match-query-types